requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM Integration
ollama==0.1.7
openai==1.3.7  # For LM Studio compatibility
llama-cpp-python==0.2.20

# MQTT
paho-mqtt==1.6.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
requests==2.31.0

# Web UI
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code quality
black==23.11.0
isort==5.12.0
mypy==1.7.1




